Add splash screen icon
