+++ .gitignore
@@ -28,3 +28,5 @@ doc/api/
 
 .vscode
 .idea
+
+.DS_Store
